feat: add disable code download annotation during install
