+++ src/commands/app/install.js
@@ -18,6 +18,8 @@ const fs = require('fs-extra')
 const execa = require('execa')
 const unzipper = require('unzipper')
 const { validateJsonWithSchema } = require('../../lib/install-helper')
+const { loadConfigFile, writeFile } = require('../../lib/import-helper')
+const { getObjectValue } = require('../../lib/app-helper')
 const jsYaml = require('js-yaml')
 const { USER_CONFIG_FILE, DEPLOY_CONFIG_FILE } = require('../../lib/defaults')
 const ora = require('ora')
@@ -28,6 +30,7 @@ class InstallCommand extends BaseCommand {
     const { args, flags } = await this.parse(InstallCommand)
 
     this.preRelease()
+    const appConfig = this.getFullConfig()
 
     aioLogger.debug(`flags: ${JSON.stringify(flags, null, 2)}`)
     aioLogger.debug(`args: ${JSON.stringify(args, null, 2)}`)
@@ -49,6 +52,7 @@ class InstallCommand extends BaseCommand {
     try {
       await this.validateZipDirectoryStructure(args.path)
       await this.unzipFile(args.path, outputPath)
+      await this.addCodeDownloadAnnotation(outputPath, appConfig)
       await this.validateConfig(outputPath, USER_CONFIG_FILE)
       await this.validateConfig(outputPath, DEPLOY_CONFIG_FILE)
       await this.npmInstall(flags.verbose)
@@ -138,6 +142,67 @@ class InstallCommand extends BaseCommand {
         throw new Error('App tests failed')
       }
     })
+  }
+
+  /**
+   * An annotation called code-download will be added to all actions in app.config.yaml
+   * (and linked yaml configs for example in extensions). This value will be set to false.
+   * The annotation will by default be true if not set.
+   *
+   * @param {object} outputPath the path to the app package
+   * @param {object} appConfig the app's configuration file
+   */
+  async addCodeDownloadAnnotation (outputPath, appConfig) {
+    this.spinner.start('Adding code-download annotations...')
+    // get each annotation key relative to the file it is defined in
+    /// iterate only over extensions that have actions defined
+    const fileToAnnotationKey = {}
+    Object.entries(appConfig.all)
+      .filter(([_, extConf]) => extConf.manifest?.full?.packages)
+      .forEach(([ext, extConf]) => {
+        Object.entries(extConf.manifest.full.packages)
+          .filter(([pkg, pkgConf]) => pkgConf.actions)
+          .forEach(([pkg, pkgConf]) => {
+            Object.entries(pkgConf.actions).forEach(([action, actionConf]) => {
+              const baseFullKey = ext === 'application'
+                ? `application.runtimeManifest.packages.${pkg}.actions.${action}`
+                : `extensions.${ext}.runtimeManifest.packages.${pkg}.actions.${action}`
+
+              let index
+              if (actionConf.annotations) {
+                index = appConfig.includeIndex[`${baseFullKey}.annotations`]
+              } else {
+                // the annotation object is not defined, take the parent key
+                index = appConfig.includeIndex[baseFullKey]
+              }
+              if (!fileToAnnotationKey[index.file]) {
+                fileToAnnotationKey[index.file] = []
+              }
+              fileToAnnotationKey[index.file].push(index.key) // index.key is relative to the file
+            })
+          })
+      })
+
+    // rewrite config files
+    for (const [file, keys] of Object.entries(fileToAnnotationKey)) {
+      const configFilePath = path.join(outputPath, file)
+      const { values } = loadConfigFile(configFilePath)
+
+      keys.forEach(key => {
+        const object = getObjectValue(values, key)
+        if (key.endsWith('.annotations') || key === 'annotations') {
+          // object is the annotations object
+          object['code-download'] = false
+        } else {
+          // annotation object is not defined, the object is the action object
+          object.annotations = { 'code-download': false }
+        }
+      })
+
+      // write back the modified manifest to disk
+      await writeFile(configFilePath, jsYaml.dump(values), { overwrite: true })
+    }
+    this.spinner.succeed('Added code-download annotations')
   }
 }
