+++ test/__fixtures__/install/1.all.config.json
@@ -1,0 +1,145 @@
+{
+  "all": {
+    "dx/excshell/1": {
+      "app": {
+        "hasBackend": true,
+        "hasFrontend": true,
+        "dist": "dist\\dx-excshell-1",
+        "defaultHostname": "adobeio-static.net",
+        "hostname": "adobeio-static.net",
+        "htmlCacheDuration": "60",
+        "jsCacheDuration": "604800",
+        "cssCacheDuration": "604800",
+        "imageCacheDuration": "604800",
+        "name": "my-app",
+        "version": "0.0.1"
+      },
+      "manifest": {
+        "src": "manifest.yml",
+        "full": {
+          "packages": {
+            "dx-excshell-1": {
+              "license": "Apache-2.0",
+              "actions": {
+                "generic": {
+                  "function": "src\\dx-excshell-1\\actions\\generic\\index.js",
+                  "web": "yes",
+                  "runtime": "nodejs:16",
+                  "inputs": {
+                    "LOG_LEVEL": "debug"
+                  },
+                  "annotations": {
+                    "require-adobe-auth": true,
+                    "final": true
+                  }
+                }
+              }
+            }
+          }
+        },
+        "packagePlaceholder": "__APP_PACKAGE__"
+      }
+    }
+  },
+  "implements": [
+    "dx/excshell/1"
+  ],
+  "includeIndex": {
+    "extensions": {
+      "file": "app.config.yaml",
+      "key": "extensions"
+    },
+    "extensions.dx/excshell/1": {
+      "file": "app.config.yaml",
+      "key": "extensions.dx/excshell/1"
+    },
+    "extensions.dx/excshell/1.$include": {
+      "file": "app.config.yaml",
+      "key": "extensions.dx/excshell/1.$include"
+    },
+    "extensions.dx/excshell/1.runtimeManifest": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "runtimeManifest"
+    },
+    "extensions.dx/excshell/1.runtimeManifest.packages": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "runtimeManifest.packages"
+    },
+    "extensions.dx/excshell/1.runtimeManifest.packages.dx-excshell-1": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "runtimeManifest.packages.dx-excshell-1"
+    },
+    "extensions.dx/excshell/1.runtimeManifest.packages.dx-excshell-1.actions": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "runtimeManifest.packages.dx-excshell-1.actions"
+    },
+    "extensions.dx/excshell/1.runtimeManifest.packages.dx-excshell-1.actions.generic": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "runtimeManifest.packages.dx-excshell-1.actions.generic"
+    },
+    "extensions.dx/excshell/1.runtimeManifest.packages.dx-excshell-1.actions.generic.annotations": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "runtimeManifest.packages.dx-excshell-1.actions.generic.annotations"
+    },
+    "extensions.dx/excshell/1.runtimeManifest.packages.dx-excshell-1.actions.generic.annotations.final": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "runtimeManifest.packages.dx-excshell-1.actions.generic.annotations.final"
+    },
+    "extensions.dx/excshell/1.runtimeManifest.packages.dx-excshell-1.actions.generic.annotations.require-adobe-auth": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "runtimeManifest.packages.dx-excshell-1.actions.generic.annotations.require-adobe-auth"
+    },
+    "extensions.dx/excshell/1.runtimeManifest.packages.dx-excshell-1.actions.generic.inputs": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "runtimeManifest.packages.dx-excshell-1.actions.generic.inputs"
+    },
+    "extensions.dx/excshell/1.runtimeManifest.packages.dx-excshell-1.actions.generic.inputs.LOG_LEVEL": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "runtimeManifest.packages.dx-excshell-1.actions.generic.inputs.LOG_LEVEL"
+    },
+    "extensions.dx/excshell/1.runtimeManifest.packages.dx-excshell-1.actions.generic.runtime": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "runtimeManifest.packages.dx-excshell-1.actions.generic.runtime"
+    },
+    "extensions.dx/excshell/1.runtimeManifest.packages.dx-excshell-1.actions.generic.web": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "runtimeManifest.packages.dx-excshell-1.actions.generic.web"
+    },
+    "extensions.dx/excshell/1.runtimeManifest.packages.dx-excshell-1.actions.generic.function": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "runtimeManifest.packages.dx-excshell-1.actions.generic.function"
+    },
+    "extensions.dx/excshell/1.runtimeManifest.packages.dx-excshell-1.license": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "runtimeManifest.packages.dx-excshell-1.license"
+    },
+    "extensions.dx/excshell/1.web": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "web"
+    },
+    "extensions.dx/excshell/1.actions": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "actions"
+    },
+    "extensions.dx/excshell/1.operations": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "operations"
+    },
+    "extensions.dx/excshell/1.operations.view": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "operations.view"
+    },
+    "extensions.dx/excshell/1.operations.view.0": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "operations.view.0"
+    },
+    "extensions.dx/excshell/1.operations.view.0.impl": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "operations.view.0.impl"
+    },
+    "extensions.dx/excshell/1.operations.view.0.type": {
+      "file": "src/dx-excshell-1/ext.config.yaml",
+      "key": "operations.view.0.type"
+    }
+  }
+}
