+++ test/__fixtures__/install/1.annotation-added.config.json
@@ -1,0 +1,34 @@
+{
+  "operations": {
+    "view": [
+      {
+        "type": "web",
+        "impl": "index.html"
+      }
+    ]
+  },
+  "actions": "actions",
+  "web": "web-src",
+  "runtimeManifest": {
+    "packages": {
+      "dx-excshell-1": {
+        "license": "Apache-2.0",
+        "actions": {
+          "generic": {
+            "function": "actions/generic/index.js",
+            "web": "yes",
+            "runtime": "nodejs:16",
+            "inputs": {
+              "LOG_LEVEL": "debug"
+            },
+            "annotations": {
+              "require-adobe-auth": true,
+              "final": true,
+              "code-download": false
+            }
+          }
+        }
+      }
+    }
+  }
+}
