+++ test/__fixtures__/install/5.app.config-loaded.json
@@ -1,0 +1,18 @@
+{
+  "values": {
+    "application": {
+      "runtimeManifest": {
+        "packages": {
+          "mypackage": {
+            "actions": {
+              "generic3": {
+                "function": "actions/generic/index.js",
+                "runtime": "nodejs:16"
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
