+++ test/__fixtures__/install/5.sub1.annotation-added.config.json
@@ -1,0 +1,22 @@
+{
+  "generic": {
+    "function": "actions/generic/index.js",
+    "web": "yes",
+    "runtime": "nodejs:16",
+    "inputs": {
+      "LOG_LEVEL": "debug"
+    },
+    "annotations": {
+      "require-adobe-auth": true,
+      "final": true,
+      "code-download": false
+    }
+  },
+  "generic4": {
+    "function": "actions/generic/index.js",
+    "runtime": "nodejs:16",
+    "annotations": {
+      "code-download": false
+    }
+  }
+}
