+++ test/__fixtures__/install/5.sub1.config-loaded.json
@@ -1,0 +1,21 @@
+{
+  "values": {
+    "generic": {
+      "function": "actions/generic/index.js",
+      "web": "yes",
+      "runtime": "nodejs:16",
+      "inputs": {
+        "LOG_LEVEL": "debug"
+      },
+      "annotations": {
+        "require-adobe-auth": true,
+        "final": true
+      }
+    },
+    "generic4": {
+      "function": "actions/generic/index.js",
+      "runtime": "nodejs:16",
+      "annotations": {}
+    }
+  }
+}
