+++ test/__fixtures__/install/5.sub2.annotation-added.config.json
@@ -1,0 +1,5 @@
+{
+  "annotations": {
+    "code-download": false
+  }
+}
