+++ test/__fixtures__/install/5.sub2.config-loaded.json
@@ -1,0 +1,3 @@
+{
+  "values": { "annotations": {} }
+}
