+++ test/commands/app/install.test.js
@@ -16,6 +16,7 @@ const fs = require('fs-extra')
 const unzipper = require('unzipper')
 const execa = require('execa')
 const installHelper = require('../../../src/lib/install-helper')
+const importHelper = require('../../../src/lib/import-helper')
 const { USER_CONFIG_FILE, DEPLOY_CONFIG_FILE } = require('../../../src/lib/defaults')
 const path = require('node:path')
 const jsYaml = require('js-yaml')
@@ -23,9 +24,16 @@ const jsYaml = require('js-yaml')
 jest.mock('fs-extra')
 jest.mock('unzipper')
 jest.mock('../../../src/lib/install-helper')
+jest.mock('../../../src/lib/import-helper')
 jest.mock('js-yaml')
 jest.mock('ora')
 jest.mock('execa')
+
+const mockGetFullConfig = jest.fn()
+
+beforeAll(() => {
+  jest.spyOn(BaseCommand.prototype, 'getFullConfig').mockImplementation(mockGetFullConfig)
+})
 
 const mockReadStreamPipe = jest.fn()
 const mockUnzipExtract = jest.fn()
@@ -73,6 +81,7 @@ beforeEach(() => {
   process.cwd = jest.fn().mockImplementation(() => fakeCwd)
   process.chdir.mockClear()
   process.cwd.mockClear()
+  mockGetFullConfig.mockClear()
 })
 
 test('exports', () => {
@@ -285,6 +294,7 @@ describe('run', () => {
     // since we already unit test the methods above, we mock it here
     command.validateZipDirectoryStructure = jest.fn()
     command.unzipFile = jest.fn()
+    command.addCodeDownloadAnnotation = jest.fn()
     command.validateConfig = jest.fn()
     command.runTests = jest.fn()
     command.npmInstall = jest.fn()
@@ -309,6 +319,7 @@ describe('run', () => {
     // we only reject one call, to simulate a subcommand failure
     command.validateZipDirectoryStructure = jest.fn()
     command.unzipFile = jest.fn()
+    command.addCodeDownloadAnnotation = jest.fn()
     command.validateConfig = jest.fn()
     command.npmInstall = jest.fn()
     command.error = jest.fn()
@@ -336,6 +347,7 @@ describe('run', () => {
     // we only reject one call, to simulate a subcommand failure
     command.validateZipDirectoryStructure = jest.fn()
     command.unzipFile = jest.fn()
+    command.addCodeDownloadAnnotation = jest.fn()
     command.validateConfig = jest.fn()
     command.npmInstall = jest.fn()
     command.error = jest.fn()
@@ -360,6 +372,7 @@ describe('run', () => {
     // since we already unit test the methods above, we mock it here
     command.validateZipDirectoryStructure = jest.fn()
     command.unzipFile = jest.fn()
+    command.addCodeDownloadAnnotation = jest.fn()
     command.validateConfig = jest.fn()
     command.runTests = jest.fn()
     command.npmInstall = jest.fn()
@@ -375,4 +388,82 @@ describe('run', () => {
     expect(command.error).toHaveBeenCalledTimes(0)
     expect(fakeCwd).toEqual(path.resolve('my-dest-folder'))
   })
+})
+
+test('addCodeDownloadAnnotation: default', async () => {
+  const extConfig = fixtureJson('install/1.all.config.json')
+
+  importHelper.loadConfigFile.mockImplementation(() => {
+    return fixtureJson('install/1.ext.config-loaded.json')
+  })
+
+  const command = new TheCommand()
+  command.argv = []
+  await command.addCodeDownloadAnnotation('my-dest-folder', extConfig)
+
+  expect(importHelper.writeFile).toHaveBeenCalledWith(
+    path.join('my-dest-folder', 'src', 'dx-excshell-1', 'ext.config.yaml'),
+    jsYaml.dump(fixtureJson('install/1.annotation-added.config.json')),
+    { overwrite: true }
+  )
+})
+
+test('addCodeDownloadAnnotation: no annotations defined', async () => {
+  const extConfig = fixtureJson('install/1.all.config.json')
+  // should not have any annotations set
+  delete extConfig.all['dx/excshell/1'].manifest.full.packages['dx-excshell-1'].actions.generic.annotations
+
+  const fixtureLoaded = fixtureJson('install/1.ext.config-loaded.json')
+  delete fixtureLoaded.values.runtimeManifest.packages['dx-excshell-1'].actions.generic.annotations
+  const fixtureExpected = fixtureJson('install/1.annotation-added.config.json')
+  fixtureExpected.runtimeManifest.packages['dx-excshell-1'].actions.generic.annotations = {
+    'code-download': false
+  }
+
+  importHelper.loadConfigFile.mockReturnValue(fixtureLoaded)
+
+  const command = new TheCommand()
+  command.argv = []
+  await command.addCodeDownloadAnnotation('my-dest-folder', extConfig)
+
+  expect(importHelper.writeFile).toHaveBeenCalledWith(
+    path.join('my-dest-folder', 'src', 'dx-excshell-1', 'ext.config.yaml'),
+    jsYaml.dump(fixtureExpected),
+    { overwrite: true }
+  )
+})
+
+test('addCodeDownloadAnnotation: complex includes, multiple actions and extensions', async () => {
+  const extConfig = fixtureJson('install/5.all.config.json')
+
+  importHelper.loadConfigFile.mockImplementation(file => {
+    const retValues = {
+      [path.join('my-dest-folder', 'app.config.yaml')]: fixtureJson('install/5.app.config-loaded.json'),
+      [path.join('my-dest-folder', 'sub1.config.yaml')]: fixtureJson('install/5.sub1.config-loaded.json'),
+      [path.join('my-dest-folder', 'src', 'sub2.config.yaml')]: fixtureJson('install/5.sub2.config-loaded.json')
+    }
+    return retValues[file]
+  })
+
+  const command = new TheCommand()
+  command.argv = []
+  await command.addCodeDownloadAnnotation('my-dest-folder', extConfig)
+
+  expect(importHelper.writeFile).toHaveBeenCalledWith(
+    path.join('my-dest-folder', 'app.config.yaml'),
+    jsYaml.dump(fixtureJson('install/5.app.annotation-added.config.json')),
+    { overwrite: true }
+  )
+
+  expect(importHelper.writeFile).toHaveBeenCalledWith(
+    path.join('my-dest-folder', 'sub1.config.yaml'),
+    jsYaml.dump(fixtureJson('install/5.sub1.annotation-added.config.json')),
+    { overwrite: true }
+  )
+
+  expect(importHelper.writeFile).toHaveBeenCalledWith(
+    path.join('my-dest-folder', 'src', 'sub2.config.yaml'),
+    jsYaml.dump(fixtureJson('install/5.sub2.annotation-added.config.json')),
+    { overwrite: true }
+  )
 })
